refactor: extract formatTable/formatRelativeTime from list.ts to shared format.ts
The prompts list command had inline table formatting. Extract to
cli/utils/format.ts (already created for dataset commands) and update
the call to use the colorMap parameter.
